requirements.txt
python-telegram-bot==20.8
aiohttp
python-dotenv
PyGithub
transformers
torch
